Add support for SPA and MVC module development
Essentially the only thing that changes in the DNN manifest for SPA and MVC modules in DNN 8 are the controlSrc tags.  ASCX files are no longer the only options.  See the following for reference:

##SPA##
* https://github.com/jbrinkman/DNN8Templates/blob/master/src/DnnSpaModule/DnnSpaModule.dnn

##MVC##
* https://github.com/jbrinkman/DNN8Templates/blob/master/src/DnnMvcModule/DnnMvcModule.dnn

As a result of this change, you may want to update your Getting Started blog entry.